+++ src/test/kotlin/KtConfigNegativeCompilationTest.kt
@@ -136,6 +136,7 @@ class KtConfigNegativeCompilationTest {
                 .withArguments(
                     "--project-cache-dir",
                     projectDir.resolve("cache").toString(),
+                    "-Dktconfig.testJarDir=${projectDir.resolve("included-jars").invariantSeparatorsPath()}",
                     "compileKotlin",
                 ).buildAndFail()
         }